fixed float index issue in unpacking data

+++ your/psrfits.py
@@ -141,7 +141,7 @@ class PsrfitsFile(object):
             if (shp[0] != self.nsamp_per_subint) and \
                     (shp[1] != self.nchan * self.nbits / 8):
                 sdata = sdata.reshape(self.nsamp_per_subint,
-                                      self.nchan * self.nbits / 8)
+                                      int(self.nchan * self.nbits / 8))
             if self.nbits == 4:
                 data = unpack_4bit(sdata)
             elif self.nbits == 2:
